feat(ui): Add final updates and screenshots. Closes #7
Add screenshots
Add fixtures of tickets, labels and status. Modify gitignore for fixtures to be tracked
Add static files
Update UpdateAccountView. Make it work with a custom Form instead of the default ModelForm created by UpdateView.
Add the `change_all_dates` method to Ticket model. This method is to be used for showcasing the application properly by adjusting dates to be displayed on the dashboard.
Update DashboardView for filtering only tickets published in the past (in case a use of `change_all_dates` brings a pub_date to the future)